libs/port: add ctzl and ffsl, minor optimization to ffs
ctzl is needed for efficient bitmap operations and is available from GCC
as a built-in (via ctol) since 3.3.4.  When not available, a C version
of ffsl is used instead.

The optimization to ffs is only very slightly faster (tested on phenom),
but does reduce text size a little bit as well.

Signed-off-by: Daniel Santos <daniel.santos@pobox.com>