libs/port: move small_pause into port.h and rename
I've renamed it cpu_relax() as, as I understand it, rep; nop; doesn't
actually pause so much as tells the CPU to dip into a lower power mode
temporarily. Either way, this will create a memory barrier which I hear
takes about 100 cycles on the average x86(_64) CPU.

This function is needed for hybrid sync stuff.

Signed-off-by: Daniel Santos <daniel.santos@pobox.com>