server: add shared memory slab allocator
An efficient slab allocator for shared memory.  This is used to allocate
the shared portion of hybrid objects in shared memory and manage the
details that another process will need to access that memory while and
preventing waste of memory and file descriptors.

The allocator performs some level of memory debugging to catch misuse by
(optionally) padding objects, poisoning memory and verifying that memory
that is either padding or unallocated is untouched. These features are
optimized by the fact that it works with memory in "nibblets" (type
nibblet_t) which are pretty much either 4 or 8 bytes depending upon what
the CPU manages more quickly.

Signed-off-by: Daniel Santos <daniel.santos@pobox.com>