server: add pretty_dump
This is a stupid object dumper that helps in debugging and can be
removed when development of this patchset is finished.

Alternatively, it can be improved and used as a replacement for all code
where objects are dumped. Specifically, it would be nice if this could
just have a dump_field function and depending upon how you create the
struct pretty_dump object, it can either print them with all field names
and values aligned or it could print them compactly (like wineserver
does now). As such, it could also be modified a bit further to print
json, but it's very likely that
https://en.wikipedia.org/wiki/Wikipedia:Nobody_cares.

Signed-off-by: Daniel Santos <daniel.santos@pobox.com>