server: add get_handle_hybrid_obj (analogy to get_handle_obj)
This is the analogy of get_handle_obj() for retrieving hybrid objects.
Currently, an object is considered hybrid if
obj->ops->trywait_begin_trans is non-null, but this is crappy. The
mechanism is basically bolted on to the side of struct object's
polymorphism system. To clean this up, we probably need to alter how
struct object does polymorphism (perhaps a type id?) because a "hybrid
object" is an essentially an abstract type, like "object".

None the less, this function cleanly returns the object if it is hybrid
and fails correctly if not (without needing to pass an ops pointer).

Signed-off-by: Daniel Santos <daniel.santos@pobox.com>