server: add process group management
Manages the relationship between hybrid sync obects and the processes
that have access to them and ensures that objects are only mapped into
the memory space of objects that have live handles to them.

This code appears to still have a flaw in it somewhere that only
expresses when running the migration stress test with 32+ thread/process
pairs.

Signed-off-by: Daniel Santos <daniel.santos@pobox.com>